Almost complete a few more tweaks and new release will be created
