+++ appLauncher/Core/Pages/settings.xaml
