+++ appLauncher/Core/Pages/settings.xaml.cs
